+++ dev-packages/cli-service/src/serve/serve-manager.ts
@@ -4,7 +4,7 @@ import { HookExecutor } from '../hooks';
 import * as https from 'https';
 import * as http from 'http';
 import { ConfigurationContext, ServiceContextUtils } from '../context';
-import { LoggerUtil, spawnProcess } from '@malagu/cli-common';
+import { ConfigUtil, LoggerUtil, spawnProcess } from '@malagu/cli-common';
 
 export type ExecuteServeHooks = (server: http.Server | https.Server, app: Express.Application,
     compiler: webpack.Compiler, entryContextProvider: () => Promise<any>) => Promise<void>;
@@ -25,13 +25,16 @@ export class ServeManager {
     async start() {
         this.log();
 
-        const compileCommand: string = this.context.framework?.settings?.compileCommand;
+        const backendConfig = ConfigUtil.getBackendConfig(this.context.cfg);
+        const frontendConfig = ConfigUtil.getFrontendConfig(this.context.cfg);
+
+        const compileCommand: string = backendConfig.compileCommand || frontendConfig.compileCommand;
         if (compileCommand) {
             const args = compileCommand.split(/\s+/);
             await spawnProcess(args.shift()!, args, { stdio: 'inherit' });
         }
 
-        const serveCommand: string = this.context.framework?.settings?.serveCommand;
+        const serveCommand: string = backendConfig.serveCommand || frontendConfig.serveCommand;
         if (serveCommand) {
             const args = serveCommand.split(/\s+/);
             await spawnProcess(args.shift()!, args, { stdio: 'inherit' });