+++ plugins/static-plugin/src/hooks/webpack.ts
@@ -1,15 +1,16 @@
-import { ConfigurationContext, WebpackContext, PathUtil } from '@malagu/cli-service';
-import { join } from 'path';
+import { ConfigurationContext, WebpackContext, PathUtil, ConfigUtil } from '@malagu/cli-service';
+import { resolve } from 'path';
 const rimraf = require('rimraf');
 
 export default async (context: WebpackContext) => {
-    const { configurations, framework, runtime } = context;
+    const { configurations, runtime, cfg } = context;
     const config = ConfigurationContext.getBackendConfiguration(
         configurations
     );
     const projectHomePath = PathUtil.getProjectHomePath(runtime);
     if (config && !projectHomePath.startsWith(process.cwd())) {
-        if (framework) {
+        const outputDir = ConfigUtil.getFrontendConfig(cfg).outputDir;
+        if (outputDir) {
             const CopyPlugin = require('copy-webpack-plugin');
             const to = PathUtil.getFrontendProjectDistPath(runtime);
             rimraf.sync(to);
@@ -17,7 +18,7 @@ export default async (context: WebpackContext) => {
                 .plugin('copyFrontend')
                 .use(CopyPlugin, [{
                     patterns: [{
-                        from: join(process.cwd(), framework.settings?.outputDir || 'dist'),
+                        from: resolve(process.cwd(), outputDir),
                         to
                     }]
                 }]);